Fixed Nvidia GPU isolator build for gcc.
There appears to be a discrepancy between clang and gcc, which allows
clang to accept `using` declarations of the form `using ns_name::name;`
that contain nested classes, structs, and enums after the `name` field
in the declaration (e.g. `using ns_name::name::enum;`).

The language for describing this functionality is ambiguous in the
C++11 specification as referenced here:
http://en.cppreference.com/w/cpp/language/namespace#Using-declarations

This patch fixes a bug where we relied on this functionality in clang
and had build errors in gcc.

Review: https://reviews.apache.org/r/45622/