Fixed access to /dev/nvidia{ctl,-uvm} in Nvidia GPU isolator.
Previously, calls to 'nvidia-smi' would fail inside a container even
if access to a GPU had been granted. Moreover, access to
/dev/nvidiactl is actually required for a container to do anything
useful with a GPU even if it has access to it.

This patch explicitly grants/revokes access to /dev/nvidiactl and
/dev/nvidia-uvm as GPUs are added and removed from a container in
the Nvidia GPU isolator.

Review: https://reviews.apache.org/r/45715/