Reset the `LIBPROCESS_IP` environment variable in `network/cni`.
Currently the `LIBPROCESS_IP` environment variable was being set to the
Agent IP if the environment variable has not be defined by the
`Framework`. For containers having their own IP address (as with
containers on CNI networks) this becomes a problem since the command
executor tries to bind to the `LIBPROCESS_IP` that does not exist in its
network namespace, and fails. Thus, for containers launched on CNI
networks the `LIBPROCESS_IP` should not be set, or rather is set to
"0.0.0.0", allowing the container to bind to the IP address provided by
the CNI network.

Review: https://reviews.apache.org/r/45801/