Fixed a parsing bug in the hashmap flags parser.
We leverage the JSON format to parse maps, but we do not ensure that
values are strings, so we end up tolerating input inconsistently:

(1) {"k", {}}   // Valid
(2) {"k", "{}"} // Valid
(3) {k, "{}"}   // Invalid

It seems inconsistent that (1) is accepted and (3) is not. Also,
perhaps counter-intuitively, (1) and (2) yield the same result.

After this change, only (2) is considered valid.

Review: https://reviews.apache.org/r/44832/