Made `Delegate` and `Handlers` libprocess tests less fragile.
As previously written, these tests depended on the fact that `post`
will synchronously deliver a message to a local PID.  That is an
implementation detail that seems unwise to rely upon. Instead, it is
quite easy to arrange for both tests to block until the effect of
the `post` has occurred.

Review: https://reviews.apache.org/r/45590/