Windows: Add Windows-friendly implementation of `rm.hpp`.
We currently depend on arcane "POSIX-like" Windows APIs like `::remove`.
This API is incompatible with NT paths -- which we will eventually have
to transition to for the Windows integration -- and has no documented
behavior interacting with subsystems like the driver subsystem, which is
required for symlinks.

This commit will move us to proper core Windows API equivalents for
`::remove`.

Review: https://reviews.apache.org/r/45022/