Made sure all Python modules are installed.
The way our build sets up PYTHONPATH it appears to pip as if the
mesos.cli package is already installed when it is only present in some
directory in PYTHONPATH, but not in the installation target location.
pip subsequently skips mesos.cli leading to a broken install where
files required by e.g., mesos-ps are not installed.

Explicitly ignoring installed packages lets us install mesos.cli.

This appears to be due to some change in behavior first occurring with
pip-6.0.1.

Review: https://reviews.apache.org/r/45209/