Fixed per framework principal metrics for HTTP frameworks.
Use the `principal` set in `FrameworkInfo.principal` to populate
framework metrics instead of looking them up in `authenticated`
map. It also ensures that HTTP->PID framework downgrades preserve
the metrics and no additional logic is needed to deal with this.
(Previously HTTP frameworks never had AuthN and no principal was
passed in \`FrameworkInfo\`, so this was never a concern).

Review: https://reviews.apache.org/r/46114/