Add subdirectory support to URI.output_file field.
URI.output_file allows the user to specify the path of the file that'll
be saved in the sandbox when the URI is fetched, but previously it would
fail at fetch time if "filename" had a directory component. This change
allows users to specify a relative path for custom ouput targets within
the sandbox.

The protobuf field 'filename' is renamed 'output_file' due to this
usage change.

Review: https://reviews.apache.org/r/46168/