Added helper to create test agent with injected `Authorizer`.
In addition to the fully generic interface we do provide a number of
short hand functions for creating agents in tests which allow injecting
just a single component. Add one such short hand function for creating
a test agent with an injected `Authorizer` which we will use in a
subsequent patch.

Review: https://reviews.apache.org/r/46318/