Send error message to the framework when it is not connected.
This change adds logic to send the framework an error message
if the master receives a call from the framework after it has
disconnected/when the master notices a framework disconnection.
Since driver based frameworks don't receive heartbeats from the
master. There is no way for them to detect this unless a new
master is detected that eventually triggers re-registration.

Review: https://reviews.apache.org/r/46723/