Patched glog to eliminate system-level gflag library detection.
The glog library will attempt to detect whether the gflag library
is installed on the system. However, glog can compile correctly
when gflags is absent, due to a compatiblity layer:

https://github.com/google/glog/blob/v0.3.3/src/base/commandlineflags.h

We do not want a 3rd party dependency on the host to be picked up
by glog in general. In the case of gflags, it turns out that this
may break the build (when the gflag headers are incompatible).

There is no configure flag for disabling this auto-detection, so
we patch glog to remove it.

Review: https://reviews.apache.org/r/46804