Skipped the bind mounting of the agent's work_dir if possible.
In filesystem/linux isolator, we will need to make sure the mount that
contains the agent's working directory is a shared mount in its own
peer group so that cloning the host mount table (e.g., when launching
a container with a new mount namespace) does not create an extra
reference to the mounts under the agent's work_dir.

However, if the mount containing the agent's working directory is
already a shared mount in its own peer group, we don't have to create
another bind mount for the agent's working directory. This patch
allows the filesystem/linux isolator to skip the bind mount if
possible.

Review: https://reviews.apache.org/r/46858