Updated HTTP command executor to support kill policy in Kill event.
Kill policy can be provided in a kill event. In this case it should
take precedence over kill policy specified when the task was launched.
When kill event is issued multiple times during the task termination,
the signal escalation timeout (the time a task has between SIGTERM
and SIGKILL) may be reduced.

Since updating the delay timer (we use it for signal escalation delay)
is currently not possible, we cancel the existing signal timer and set
up a new one. `Clock::cancel()` guarantees that, if existed, the timer
is removed before the function returns; hence we do not set up more
than 1 timer for signal escalation delay.

Review: https://reviews.apache.org/r/46325/