Updated agent to backoff before linking with the master.
Currently the agent will immediately link with the master once
the master is discovered. The agent currently uses a registration
backoff in order to alleviate the failover load of the master for
large clusters. However, since the link occurs immediately, the
master still has to process a large influx of connections. This
can lead to connections closing and the agents not re-registering
due to MESOS-1963.

Performing the link after the initial backoff ensures that
the --registration_backoff flag alleviates a thundering herd
of TCP connections against the master.

Review: https://reviews.apache.org/r/47209/