Moved code from HTTP command executor to command executor.
This is a blatant copy of `src/launcher/http_command_executor.cpp`.
This change makes reviewing code later in the chain easier.
Later in the chain, we would make this use the Shim/Adapter
to toggle between using the driver or the v1 API via an
environment variable. The existing \`http_command_executor.cpp\`
would be deleted.

Review: https://reviews.apache.org/r/47365/