Rewrote os::read() to be friendlier to reading binary data.
The previous read() implementation was based on calling getline() to
read in chunks of data from a file. This is fine for text-based files,
but is a little strange for binary files.

The new implementation reads in chunks of raw bytes into a buffer
before copying them into their final location.

Review: https://reviews.apache.org/r/47481/