Added preliminary support for parsing ELF files in stout.
Right now we are able to parse ELF formatted shared libraries and
extract their canonical SONAME and external library dependencies. In
the future, we should add support for fully parsing an ELf file for
easy access to all of its contents.

The current implementation relies on libelf. We should probably remove
this dependency in future versions (mostly since the headers for
libelf are not installed on a standard Linux distribution by default).

Review: https://reviews.apache.org/r/47482/