Fixed broken libelf dependency.
Unfortunately, we recently introduced a global dependency on libelf,
which broke the mesos build. The only subsystem that actually depends
on libelf is the Nvidia GPU isolator, so we should really have only
been adding the dependency in cases where ENABLE_NVIDIA_GPU_SUPPORT
was set. This patch moves the logic to require the dependency only
when this flag is set.

Review: https://reviews.apache.org/r/47601/