Fixed a race in long lived executor.
There is a possible race condition when the \`TASK_FINISHED\`
update calls into \`update\` when the executor process might
itself be already executing. Also after this fix, even if the
executor is disconnected from the agent by the time the `update`
function is called, the updates would be sent again as part of
unacknowledged updates on re-registration.

Review: https://reviews.apache.org/r/47732/