Edited `--work_dir` configuration docs.
Some users have encountered difficulty when running the Mesos agent
with the `work_dir` located in a subdirectory of `/tmp`. This patch
adds language to the `work_dir` help strings and configuration docs
advising users to avoid the use of this location in production.

Review: https://reviews.apache.org/r/45562/