Added `ObjectApprover` interface to `Authorizer`.
With the goal to provide more efficient authorization for multiple,
potentially large objects, we extend the authorizer interface to support
`ObjectApprover`. Retrieving an `ObjectApprover` for a given action is
an asynchronous operation but following authorization of multiple
objects can be done synchronously without copying.

NOTE: This implies that `Authorizer` module writers need to ensure that
an `ObjectApprover` will behave nicely (e.g., it must not block).

Review: https://reviews.apache.org/r/47558/