Avoid copying the Request for each streaming chunk in libprocess.
This change passes the `Request` object by wrapping it in
`Owned` instead of copying it on each streaming read from the
`Pipe`. Ideally, we can make the `Request` an `Owned` or `Shared`
object from the point where it is decoded to avoid copying it
around in libprocess.

Review: https://reviews.apache.org/r/47991/