Replaced RUN_TASK_WITH_USER with RUN_TASK authorization action.
Authorization requests for RUN_TASK actions can pass `SOME`
authorization object either in a `FrameworkInfo` holding a user, or a
`TaskInfo`/`ExecutorInfo` with a `CommandInfo` holding a user.
If any of these fields is set it will be used as the object;
otherwise an `ANY` type authorization object will be created.

Review: https://reviews.apache.org/r/48113