Added configure check for separate fts library to Mesos.
On some systems (e.g., those that use musl instead of glibc),
the fts(3) family of functions are not provided by libc.
Instead, a separate fts library must be installed. If this is
the case, "libfts" should be added to the list of libraries that
stout is linked against.

Review: https://reviews.apache.org/r/48247/