Updated Docker::run to return exit status of container.
Previously, Docker::run would return a Future<Nothing> which
requires the caller to perform an inspect to obtain the exit
status of the container. This is difficult to get right, and
so this patch leverages the fact that 'docker run' already
returns the exit status from the container termination to
simplify the calling code. Namely, in the docker executor we
need to use the exit status to determine how to send a
terminal status update (see the command executor code).

Review: https://reviews.apache.org/r/48428