Remove SASL dependency for Windows builds.
SASL is currently a hard dependency for Mesos. Per MESOS-5450, we expect
that some platforms will not support SASL (namely Windows), so this
commit will begin the first step in a several-step process of removing
it as a hard dependency.

For this first step, this commit will shed SASL dependency in libmesos
only for Windows builds. We do this by:

  (1) Adding a preprocessor symbol, `HAS_AUTHENTICATION` that wraps the
      SASL-dependent code, so that we can conditionally choose not to
      compile it.
  (2) Defining `HAS_AUTHENTICATION` on all Unix builds, and leaving it
      undefined on all Windows builds.
  (3) Logging an error and exiting the master if the user passes in
      flags that depend on SASL, such as `--authenticate`.

Notably, what we do *not* do is:

  (1) Shed SASL dependency in the tests.

The impact of this is that, on Windows, relevant libmesos tests will
either not compile, or not pass. Naturally, as tracked by MESOS-5450,
the second phase of this series will be to shed the test dependency as
well.

Review: https://reviews.apache.org/r/47821/