Avoid killing a stale pid in the docker/command executors.
This is a stop-gap for issuing a kill on the stale healthPid.
However, there are additional issues with this code. Namely,
we do not reap the healthPid and so if it terminates we may
still try to kill a stale pid. Also, we do not kill the health
process upon finalizing the executor, we should consider adding
a finalize that performs the necessary cleanup. Lastly, there
are a number of other bugs and structural issues in the health
checking code, and so I've punted on doing more major changes
here in favor of at least fixing this glaring bug.

Review: https://reviews.apache.org/r/48427