Fixed broken docker logging redirection in the docker executor.
Currently, the docker executor will SIGKILL the 'docker run'
subprocess immediately upon receiving a kill task request.
This is problematic in that the 'docker run' subprocess is
responsible for streaming the stdout and stderr from the
container. When we kill the 'docker run' subprocess, we no
longer stream the output.

Review: https://reviews.apache.org/r/48429