Fixed a regression in the creation of the isolator modules.
While refactoring the isolator creation in 6fb9c024, we introduced a
regression in that isolator modules were no longer being created via
the --isolation flag values.

Also, we need to re-establish the old semantics that allowed the
operator to control the isolator ordering. A TODO has been added
for this.

Review: https://reviews.apache.org/r/48662