Restored ordering of --isolation entry processsing.
In 6fb9c024, the ordering of isolators was changed from being
specified by the operator to being specified by the code. It
turns out that the operator ordering was intended to ensure
that isolators can check their ordering requirements. If an
isolator detects that its ordering requirements are not
met, it can inform the operator to adjust the --isolation flag
via a failure message.

Review: https://reviews.apache.org/r/48671