Fixed hard-coding of Nvidia control device numbers.
Previously, the major/minor numbers of `/dev/nvidiactl` and
`/dev/nvidia-uvm` were hard-coded. This actually caused problems for
`/dev/nvidia-uvm` because its major number is part of the
"Experimental" device range on Linux.

Because this range is experimental, there is no guarantee which device
number will be assigned to it on a given machine. We actually
encountered this problem in the wild, prompting this change.

We now use `os:stat::rdev()` to extract the major/minor numbers
programatically.

Review: https://reviews.apache.org/r/48370/