Removed build-time dependency on libnvidia-ml.
This introduces an nvml:: wrapper interface that will dlopen
libnvidia-ml at run-time to avoid the build-time dependency.

Previously, libmesos depended on libnvidia-ml when built with
GPU support enabled. This meant the master and non-GPU agents
required libnvidia-ml (which was a huge pain as the library
is not easily installable if no GPUs are present!).

Review: https://reviews.apache.org/r/48364/