Fixed invalid usage of `dev_t` in NVIDIA GPU isolator.
Previously, the GPU struct specified the type of its `major` and
`minor` fields as `dev_t`, which is actually a concatenation of both
the major and minor device numbers accessible through the `major()`
and `minor()` macros. These macros return an `unsigned int` when
handed a `dev_t`, so it makes sense for these fields to be of that
type instead.

https://reviews.apache.org/r/48368/