Fixed agent switch user redundant logic in 'getExecutorInfo'.
Since the 'user' in 'FrameworkInfo' is a required field, so
'fromeworkInfo.has_user()' is always true. Currently in our semantics,
if the operator specifies the framework user as an empty string, mesos
will automagically set it as the current host user. So the slave will
always obtain the user from 'FrameworkInfo', and 'switch_user' defaults
to be true from agent flag. We would prefer not to change this semantics
because a long deprecation cycle is needed.  So in this patch, we just
fix the redundant check in slave 'getExecutorInfo' logic.

Review: https://reviews.apache.org/r/47660/