Always grant access to NVIDIA control devices.
Previously, access to `/dev/nvidiactl` and `/dev/nvidia-uvm` was only
granted/revoked from a container as GPUs were added and removed from
them. On some level, this makes sense because most jobs don't need
access to these devices unless they are also using a GPU. However,
there are cases when access to these files is appropriate, even when
not making use of a GPU. Running `nvidia-smi` to control the global
state of the underlying nvidia driver, for example.

This commit adds `/dev/nvidiactl` and `/dev/nvidia-uvm` to the default
whitelist of devices to include in every container when the
`gpu/nvidia` isolator is enabled. This allows a container to run
standard nvidia driver tools (such as `nvidia-smi`) without failing
with abnormal errors when no GPUs have been granted to it. As such,
these tools will now report that no GPUs are installed instead of
failing abnormally.

NOTE: Once we allow GPUs to be granted to containers with filesystem
isolation turned on, other criteria will be used to determine when /
if to grant access to these control devices.

https://reviews.apache.org/r/48376/