Updated `QuotaHandler` to send one authorization request per action.
As described in MESOS-5628, this patch makes `QuotaHandler` only
send one authorization request per action.

During the deprecation cycle of `ACL::SetQuota` and `ACL::RemoveQuota`,
the `LocalAuthorizer` for example still needs to know whether the action
is a set or remove. We remedy this situation by encoding type (set or
remove) by setting the `value` field in the `Object` with the string
`"SetQuota"` or `"RemoveQuota"`. With this approach, we can simply stop
setting the `value` field at the end of the deprecation cycle as opposed
to changing the public API.

Review: https://reviews.apache.org/r/48039/