Updated `LocalAuthorizer` to consolidate to the `UPDATE_QUOTA` action.
As described in https://reviews.apache.org/r/48039/, `QuotaHandler` is
updated to send only one authorization request per action.

In this patch, we update the `LocalAuthorizer` to handle `UPDATE_QUOTA`.
Note that `QuotaHandler` encodes the type of the request (set or remove)
by setting the `value` field of `Object` to `"SetQuota"` or
`"RemoveQuota"`. The `LocalAuthorizer` uses this information to support
`ACL::SetQuota` and `ACL::RemoveQuota` throughout the deprecation cycle.

Review: https://reviews.apache.org/r/48038/