Changed build to always enable Nvidia GPU support for Linux.
Previously, we supported a configure flag for
`--enable-nvidia-gpu-support` to explicitly enable Nvidia GPU support
when building Mesos on Linux. This commit removes this flag and
enables this support for all builds on Linux.

As part of this change, libelf now becomes a build dependence on Linux
for all build configurations. We update the docs and the docker build
scripts to accomodate this.

Review: https://reviews.apache.org/r/48832/