Delay before initiating a connection with master.
The scheduler library has been updated to wait before initiating a
connection with the master. The maximum amount of time waited by
the library is defined by the flag: `connection_delay_max`. For
tests, we default the delay to 0 to speed up the tests.

Review: https://reviews.apache.org/r/48387/