Removed hard dependence on 'libelf'.
We recently added a hard dependency for `libelf` on Linux. This was in
preparation for some upcoming Nvidia GPU support for injecting volumes
into containers. Since this dependence is not actually necessary for
the upcoming release, we decided to remove it for now, and rethink the
best way to add it back in later (possibly as a runtime dependence
instead of a linktime one).

For now, we just remove the set of files and tests that actually use
this functionality from the build. Namely, `ldcache.cpp` and
`ldcache_tests.cpp`. We will add them back in once we have know how we
plan to deal with this dependency.

Review: https://reviews.apache.org/r/48963/