Added check for `--isolation=gpu/nvidia` in order to autodiscover GPUs.
We add this check because we decided it makes sense to only do
autodiscovery of GPUs when this isolator is turned on (unlike for
CPUs, memory, and disk where autodiscovery happens by default). We
decided to take this approach, because GPU support is still
experimental, and is only known to work well if this isolator is
enabled. We didn't want to start advertising GPUs in our resource
offer and have people attempt to use them in scenarious we haven't
considered yet. In the future we may support other use cases, but for
now we are being cautious.

Review: https://reviews.apache.org/r/48979/