Updated DRF sorter to avoid re-sorting when not necessary.
The `dirty` bit is used to indicate that the sort order needs to be
re-calculated in the DRF sorter. We set the bit to `true` in the
appropriate cases, but we were not setting it to `false`!

This patch reset the `dirty` bit to `false` after sorting, this
ensures that only one client's share is re-calculated when operations
occur that only affect a single client.

This can also improve the performance of sorter in scenarios where
a large amount of unnecessary re-sorting is occuring.

Review: https://reviews.apache.org/r/48579/