Fix socket leak in SSL_ENABLE_DOWNGRADE code path.
When an SSL-enabled Mesos actor attempts to link, it will open an SSL
socket first.  If this fails, and downgrading to non-SSL is enabled,
the Mesos actor will create a "Poll" socket instead.

In this code path, we explicitly avoid calling `SocketManager::close`
as "closing" a link will also trigger `ExitedEvents`.  Instead, the
downgrade codepath tries to "swap" the `SocketManager`s state from the
old SSL socket, to the new Poll socket via `swap_implementing_socket`.

`swap_implementing_socket` leaks a `Socket` object, which is a
reference-counted wrapper for an FD.  Besides the memory leak, leaking
the `Socket` means that the reference count for the old SSL socket
will always be above zero.  And we only `close` the socket when the
reference count reaches zero.

Review: https://reviews.apache.org/r/49127/