Fixed an unsafe callback in the Master.
When attaching its log file, the Master passes
the `FilesProcess` an authorization callback to
use when authorizing a principal to access that
file. This callback was previously not
threadsafe, capturing the master's `this` pointer
and synchronously executing one of the master's
member functions in the wrong execution context.

This patch alters the callback to dispatch
execution of the member function to the master
process.

Review: https://reviews.apache.org/r/49132/