Fixed failed authorization of '/files/*' endpoints.
The `FilesProcess` stores an authorization callback
for every virtual path. If no callback is found when
authorizing a request, then it is assumed that path
is not authorizable, so the request is approved.
This led to a bug in which the callbacks for paths
with trailing slashes were not correctly identified,
leading to requests being incorrectly authorized.

This patch trims trailing slashes before comparing
the requested path with the authorized paths.

Review: https://reviews.apache.org/r/49131/