Added struct to share Nvidia components between containerizers.
Previously, the `NvidiaGpuAllocator` component was created directly
inside the `NvidiaGpuIsolatorProcess` even though it was designed to
be used by multiple containerizers at the same time. To allow for
simultaneous access, we created a new `NvidiaComponents` struct to hold
an `NvidiaGpuAllocator` instance and pass it to each containerizer as
it is created. This component can easily be extended to include more
cross-containerizer components later on.

Introducing the struct was helpful because it allows us to contain
the #ifdefs around the Nvidia classes that we pass around.

We create the `NvidiaComponents` instance inside
`Containerizer::create()` and pass it to both the docker containerizer
and the mesos containerizer when they are created.  The docker
containerizer currently doesn't do anything with this information, but
its interface has been updated to accomodate it. The mesos
containerizer has been updated to pass this all the way down to the
`NvidiaGpuIsolatorProcess` and exploit it.

Review: https://reviews.apache.org/r/48374/