Added the VIEW_FLAGS authorization action.
Adds authorization for viewing flags. Instead of being part of
`get_endpoints` it uses its own action `VIEW_FLAGS` which is
used to restrict access to the `/flags` endpoint, as well as
to filter the flag related results of the `/state` endpoint on
both master and agents.

Review: https://reviews.apache.org/r/49313/