Added a flag for excluding resources from fair sharing.
The intention of this flag is for the operator to specify
resources types (e.g. "cpus", "mem", "gpus", etc) that should
not be fairly shared.

The implementation of will not count the specified resources
towards a role's share of the cluster. This may be useful in
cases where our fair sharing implementation currently has
limitations. E.g. See the problem of "scarce" resources:

http://www.mail-archive.com/dev@mesos.apache.org/msg35631.html
https://issues.apache.org/jira/browse/MESOS-5377

Review: https://reviews.apache.org/r/48895/