Reimplemented ldcache_test.cpp using the new ELF abstraction in stout.
We also reenabled this test to run as part of the standard Linux test
suite. It had been disabled due to its dependence on the external
libelf library. Now that we bundle ELFIO instead of relying on this
external dependence, we can always have it enabled now.

Review: https://reviews.apache.org/r/49563/