Added a new 'NvidiaVolume' component.
This component is responsible for building up a consolidated volume of
binaries / libraries from the user-space portion of the Nvidia GPU
drivers so that it can be injected into a container as a single
volume. Its core functionality mimics that of the
'nvidia-docker-plugin': https://github.com/NVIDIA/nvidia-docker/

We currently only implement the 'create()' function which is
responsible for building up the volume itself. In a subsequent commit
we will add support for reading a Docker image manifest and deciding
whether we should inject the volume into the docker container.

Review: https://reviews.apache.org/r/49565/