Fix ProcessRemoteLinkTests that try to emulate 'stale' sockets.
The ProcessRemoteLinkTests `RemoteUseStaleLink` and
`RemoteStaleLinkRelink` were failing because the `test-linkee` was
closing sockets shortly after the test `::shutdown` the link.  This
is the expected behavior of libprocess.

Rewrites the `test-linkee` to keep all sockets open, regardless of the
state of the connection.  And re-enables the tests.

Review: https://reviews.apache.org/r/49543/