Fixed the flaky `MasterAPITest.Subscribe` test.
The test was flaky because `Clock::settle()` does not ensure
that data in flight (on the wire) is also flushed. Hence, sometimes
the update used to be sent by the master but never received by the
library before the test was completed. It would be good if libprocess
could add some logic for recognizing local HTTP events and sending
them directly as it does for messages using `send()`.

Review: https://reviews.apache.org/r/49837/