Fixed a subscriber FD leak when running tests.
This FD leak would only surface when running tests. We hold on to
a reference of the \`Connection\` object in the client so that it is
not destroyed while the connection is active. When running tests,
the IP:Port of libprocess remain the same which means the objects
keep on accumulating. In a real world cluster, we remove the
subscriber upon noticing a \_disconnection\_ i.e. this means the
socket has already been already closed upstream by Libprocess on
the server side.

Review: https://reviews.apache.org/r/49844/