Fixed the incomplete `TaskStatus` message of docker executor.
When we enable health check for the tasks run in docker container,
their `TaskStatus` messages generated by docker executor may miss
`NetworkInfo` field and agent would fill the host ip as default value.
In this change, we cache the `NetworkInfo` of the task and reuse it
when generating `TaskStatus` messages which the task's health is
updated.

Review: https://reviews.apache.org/r/50103/