Simplified the checkpointed resources handling logic.
Currently we always checkpoint resources.target and in the agent
recovery path we have to handle the target being the same as the
committed resources.info. If we don't create the target when the new
checkpointed resources are the same, then we don't need to worry about
that possibility in recovery.

Review: https://reviews.apache.org/r/50147