Fixed a memory lifetime issue in LibeventSSLSocketImp::send.
This function accidentally assumed that 'data' will live beyond the
scope of the call, by using 'data' within an asynchronous context.

This copies the data into an 'evbuffer' which will then get moved
into the output buffer in the event loop. Note that this does not
introduce an additional copy: we still have a single copy to get
'data' into the bufferevent output buffer.

Review: https://reviews.apache.org/r/50163