Speed up reservation test by lowering allocation interval.
ReserveResource test (and UnreserveResource test) in v1 operator
API is very slow, taking more than 1000 ms. If we set allocation
interval in master flag to 50ms and use that flag to start up a
master to use in the test, we can reduce the duration of that test
to as low as 120 ms. So, in this patch we made that change.

Review: https://reviews.apache.org/r/50072/