Disabled the `--registry_strict` master flag.
This flag was always marked as experimental. This flag controlled
whether the master will behave the same, regardless of whether the
master has failed over. Notably, in Mesos 1.0, enabling this flag would
mean that a partitioned agent would always be refused reregistration,
regardless of whether the master has failed over (by default, the agent
would be allowed to reregister iff the master failed over in the
interim).

For Mesos 1.1, we'd like to change the behavior so that partitioned
agents are always allowed to reregister (regardless of master
failover). To prevent developers from writing frameworks that depend on
the old, "strict" behavior (partitioned agents are always rerefused
reregistration), we'd like to disable the "--registry_strict" flag.

The code that implements the strict registry remains, as do the test
cases that depend on this behavior. However, `mesos-master` will refuse
to start if the flag has been specified.

Review: https://reviews.apache.org/r/49955/