Added example framework for testing disk quota enforcement.
This adds an example framework which exercises the behavior of the agent
when the `--enforce_container_disk_quota` agent flag is enabled.
The framework will accept a fixed amount of disk, and launch a task
that will use a configurable amount of disk.  The framework reports
metrics about tasks failed due to disk limitation.

Review: https://reviews.apache.org/r/46626/