Fixed use-after-close bug when using libevent and SSL.
The deferred call to SSL_shutdown within ~LibeventSSLSocketImpl
can occur after the socket fd has been closed by Socket::~Impl.

This can lead to a TLS Alert message being sent on any fd if
it the fd is re-used between the close and the SSL_shutdown!

A fix for this issue was put in place only for connected
sockets in ca3667f4e97e11ad30811753fdb52bc02854113f, but this
did not address accepted sockets. This patch is an attempt to
have a simpler and more efficient fix (avoid dup()).

Thanks to Jan-Philip Gehrcke for reporting the issue.

Review: https://reviews.apache.org/r/50477